backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
python-jose[cryptography]
passlib
bcrypt==4.0.1
pydantic[email]
pydantic-settings
pandas
python-dotenv
tensorflow-cpu
pillow
